jira-agent/requirements.txt
anthropic
openai
requests
python-dotenv
rich
